feat: 도메인 변수 타입 통일 진행중

+++ src/main/java/hello/fclover/controller/SellerController.java
@@ -67,7 +67,7 @@ public class SellerController {
 
         String sellerId = principal.getName();
         Seller seller = sellerService.findSellerById(sellerId);
-        BigInteger sellerNo = BigInteger.valueOf(seller.getSellerNo());
+        Long sellerNo = seller.getSellerNo();
         goods.setSellerNo(sellerNo);
         String businessNumber = seller.getBusinessNumber();
